transit relay: add a connection timeout of 60 seconds
Let us say, a user sent a file, a recipient typed in the code and then
got the transit message via wormhole and then got the offer
message. The recipient is shown the filename/size and asked for
consent. If the sender disconnects at that point or cancels or closes
the program before the content is sent, the recipient shouldn't wait
indefinitely trying to make the connection (or be at the mercy of
operating system timeouts)

The present change adds a "read/write" deadline. i.e. If Read() or
Write() takes more than the specified time (60 seconds in this
commit), then we disconnect the connection to the relay server.